+++ app/src/main/java/com/runnimal/app/android/service/impl/WalkServiceImpl.java
@@ -54,7 +54,7 @@ public class WalkServiceImpl extends AbstractService implements WalkService {
     }
 
     @Override
-    public void end(int distance, DisposableObserver<Walk> callback) {
+    public void end(float distance, DisposableObserver<Walk> callback) {
         Log.d("WALK", "end");
         if (route.size() > 1) {
             walk.setRoute(route);
